backend/requirements.txt
django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
openai==1.14.0
stripe==8.5.0
pillow==10.2.0
gunicorn==21.2.0
whitenoise==6.6.0
django-countries==7.5.1
pandas==2.2.1
faker==18.3.0
pytest-django==4.5.2
pytest-cov==4.0.0
pytest==7.3.1
djangorestframework-simplejwt>=5.3.1